move logging module to root of agent_chat
