+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/_base_team.py
@@ -5,7 +5,7 @@ from typing import List, Protocol
 from autogen_agentchat.agents._base_chat_agent import ChatMessage
 from autogen_core.application.logging import EVENT_LOGGER_NAME
 
-from .logging import ConsoleLogHandler, EVENT_LOGGER_NAME
+from ..logging import ConsoleLogHandler, EVENT_LOGGER_NAME
 
 logger = logging.getLogger(EVENT_LOGGER_NAME)
 logger.setLevel(logging.INFO)
