+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/group_chat/_base_chat_agent_container.py
@@ -11,7 +11,7 @@ from autogen_core.components.tool_agent import ToolException
 from ...agents import BaseChatAgent, MultiModalMessage, StopMessage, TextMessage, ToolCallMessage, ToolCallResultMessage
 from ._events import ContentPublishEvent, ContentRequestEvent
 from ._sequential_routed_agent import SequentialRoutedAgent
-from ..logging import EVENT_LOGGER_NAME
+from ...logging import EVENT_LOGGER_NAME
 
 
 class BaseChatAgentContainer(SequentialRoutedAgent):
